requirements.txt
Flask>=1.1.1
gunicorn>=19.9.0
numpy>=1.22.3
scikit-learn>=0.20
matplotlib>=3.5.1
pandas>=1.4.2
tensorflow>=2.8.0
keras>=2.7.0
Pillow>=9.1.0
os-sys>=2.1.4
opencv-python>=4.5.5.64